remove useless settings, set postgres environment variables in handler
