+++ lib/titiler-pgstac-api/runtime/dev_requirements.txt
@@ -1,2 +1,1 @@
 uvicorn
-pydantic[dotenv]
